+++ src/data/field_event_obj/berry_tree_graphics_tables.h
@@ -10,7 +10,7 @@ const struct SpriteFrameImage gEventObjectPicTable_PechaBerryTree[] = {
     overworld_frame(gEventObjectPic_PechaBerryTree, 2, 4, 5),
 };
 
-const u8 gBerryTreePaletteSlotTable_Pecha[] = {3, 4, 4, 4, 4};
+const u16 gBerryTreePaletteSlotTable_Pecha[] = {3, 4, 4, 4, 4};
 
 const struct SpriteFrameImage gEventObjectPicTable_KelpsyBerryTree[] = {
     overworld_frame(gEventObjectPic_BerryTreeDirtPile, 2, 2, 0),
@@ -24,7 +24,7 @@ const struct SpriteFrameImage gEventObjectPicTable_KelpsyBerryTree[] = {
     overworld_frame(gEventObjectPic_KelpsyBerryTree, 2, 4, 5),
 };
 
-const u8 gBerryTreePaletteSlotTable_Kelpsy[] = {3, 4, 2, 2, 2};
+const u16 gBerryTreePaletteSlotTable_Kelpsy[] = {3, 4, 2, 2, 2};
 
 const struct SpriteFrameImage gEventObjectPicTable_WepearBerryTree[] = {
     overworld_frame(gEventObjectPic_BerryTreeDirtPile, 2, 2, 0),
@@ -38,7 +38,7 @@ const struct SpriteFrameImage gEventObjectPicTable_WepearBerryTree[] = {
     overworld_frame(gEventObjectPic_WepearBerryTree, 2, 4, 5),
 };
 
-const u8 gBerryTreePaletteSlotTable_Wepear[] = {3, 4, 2, 2, 2};
+const u16 gBerryTreePaletteSlotTable_Wepear[] = {3, 4, 2, 2, 2};
 
 const struct SpriteFrameImage gEventObjectPicTable_IapapaBerryTree[] = {
     overworld_frame(gEventObjectPic_BerryTreeDirtPile, 2, 2, 0),
@@ -52,7 +52,7 @@ const struct SpriteFrameImage gEventObjectPicTable_IapapaBerryTree[] = {
     overworld_frame(gEventObjectPic_IapapaBerryTree, 2, 4, 5),
 };
 
-const u8 gBerryTreePaletteSlotTable_Iapapa[] = {3, 4, 3, 3, 3};
+const u16 gBerryTreePaletteSlotTable_Iapapa[] = {3, 4, 3, 3, 3};
 
 const struct SpriteFrameImage gEventObjectPicTable_CheriBerryTree[] = {
     overworld_frame(gEventObjectPic_BerryTreeDirtPile, 2, 2, 0),
@@ -66,7 +66,7 @@ const struct SpriteFrameImage gEventObjectPicTable_CheriBerryTree[] = {
     overworld_frame(gEventObjectPic_CheriBerryTree, 2, 4, 5),
 };
 
-const u8 gBerryTreePaletteSlotTable_Cheri[] = {3, 4, 4, 4, 4};
+const u16 gBerryTreePaletteSlotTable_Cheri[] = {3, 4, 4, 4, 4};
 
 const struct SpriteFrameImage gEventObjectPicTable_FigyBerryTree[] = {
     overworld_frame(gEventObjectPic_BerryTreeDirtPile, 2, 2, 0),
@@ -80,7 +80,7 @@ const struct SpriteFrameImage gEventObjectPicTable_FigyBerryTree[] = {
     overworld_frame(gEventObjectPic_FigyBerryTree, 2, 4, 5),
 };
 
-const u8 gBerryTreePaletteSlotTable_Figy[] = {3, 4, 3, 3, 3};
+const u16 gBerryTreePaletteSlotTable_Figy[] = {3, 4, 3, 3, 3};
 
 const struct SpriteFrameImage gEventObjectPicTable_MagoBerryTree[] = {
     overworld_frame(gEventObjectPic_BerryTreeDirtPile, 2, 2, 0),
@@ -94,7 +94,7 @@ const struct SpriteFrameImage gEventObjectPicTable_MagoBerryTree[] = {
     overworld_frame(gEventObjectPic_MagoBerryTree, 2, 4, 5),
 };
 
-const u8 gBerryTreePaletteSlotTable_Mago[] = {3, 4, 3, 3, 3};
+const u16 gBerryTreePaletteSlotTable_Mago[] = {3, 4, 3, 3, 3};
 
 const struct SpriteFrameImage gEventObjectPicTable_LumBerryTree[] = {
     overworld_frame(gEventObjectPic_BerryTreeDirtPile, 2, 2, 0),
@@ -108,7 +108,7 @@ const struct SpriteFrameImage gEventObjectPicTable_LumBerryTree[] = {
     overworld_frame(gEventObjectPic_LumBerryTree, 2, 4, 5),
 };
 
-const u8 gBerryTreePaletteSlotTable_Lum[] = {3, 4, 4, 4, 4};
+const u16 gBerryTreePaletteSlotTable_Lum[] = {3, 4, 4, 4, 4};
 
 const struct SpriteFrameImage gEventObjectPicTable_RazzBerryTree[] = {
     overworld_frame(gEventObjectPic_BerryTreeDirtPile, 2, 2, 0),
@@ -122,7 +122,7 @@ const struct SpriteFrameImage gEventObjectPicTable_RazzBerryTree[] = {
     overworld_frame(gEventObjectPic_RazzBerryTree, 2, 4, 5),
 };
 
-const u8 gBerryTreePaletteSlotTable_Razz[] = {3, 4, 4, 4, 4};
+const u16 gBerryTreePaletteSlotTable_Razz[] = {3, 4, 4, 4, 4};
 
 const struct SpriteFrameImage gEventObjectPicTable_GrepaBerryTree[] = {
     overworld_frame(gEventObjectPic_BerryTreeDirtPile, 2, 2, 0),
@@ -136,7 +136,7 @@ const struct SpriteFrameImage gEventObjectPicTable_GrepaBerryTree[] = {
     overworld_frame(gEventObjectPic_GrepaBerryTree, 2, 4, 5),
 };
 
-const u8 gBerryTreePaletteSlotTable_Grepa[] = {3, 4, 3, 3, 3};
+const u16 gBerryTreePaletteSlotTable_Grepa[] = {3, 4, 3, 3, 3};
 
 const struct SpriteFrameImage gEventObjectPicTable_RabutaBerryTree[] = {
     overworld_frame(gEventObjectPic_BerryTreeDirtPile, 2, 2, 0),
@@ -150,7 +150,7 @@ const struct SpriteFrameImage gEventObjectPicTable_RabutaBerryTree[] = {
     overworld_frame(gEventObjectPic_RabutaBerryTree, 2, 4, 5),
 };
 
-const u8 gBerryTreePaletteSlotTable_Rabuta[] = {3, 4, 4, 4, 4};
+const u16 gBerryTreePaletteSlotTable_Rabuta[] = {3, 4, 4, 4, 4};
 
 const struct SpriteFrameImage gEventObjectPicTable_NomelBerryTree[] = {
     overworld_frame(gEventObjectPic_BerryTreeDirtPile, 2, 2, 0),
@@ -164,7 +164,7 @@ const struct SpriteFrameImage gEventObjectPicTable_NomelBerryTree[] = {
     overworld_frame(gEventObjectPic_NomelBerryTree, 2, 4, 5),
 };
 
-const u8 gBerryTreePaletteSlotTable_Nomel[] = {3, 4, 3, 3, 3};
+const u16 gBerryTreePaletteSlotTable_Nomel[] = {3, 4, 3, 3, 3};
 
 const struct SpriteFrameImage gEventObjectPicTable_LeppaBerryTree[] = {
     overworld_frame(gEventObjectPic_BerryTreeDirtPile, 2, 2, 0),
@@ -178,7 +178,7 @@ const struct SpriteFrameImage gEventObjectPicTable_LeppaBerryTree[] = {
     overworld_frame(gEventObjectPic_LeppaBerryTree, 2, 4, 5),
 };
 
-const u8 gBerryTreePaletteSlotTable_Leppa[] = {3, 4, 3, 3, 3};
+const u16 gBerryTreePaletteSlotTable_Leppa[] = {3, 4, 3, 3, 3};
 
 const struct SpriteFrameImage gEventObjectPicTable_LiechiBerryTree[] = {
     overworld_frame(gEventObjectPic_BerryTreeDirtPile, 2, 2, 0),
@@ -192,7 +192,7 @@ const struct SpriteFrameImage gEventObjectPicTable_LiechiBerryTree[] = {
     overworld_frame(gEventObjectPic_LiechiBerryTree, 2, 4, 5),
 };
 
-const u8 gBerryTreePaletteSlotTable_Liechi[] = {3, 4, 4, 4, 4};
+const u16 gBerryTreePaletteSlotTable_Liechi[] = {3, 4, 4, 4, 4};
 
 const struct SpriteFrameImage gEventObjectPicTable_HondewBerryTree[] = {
     overworld_frame(gEventObjectPic_BerryTreeDirtPile, 2, 2, 0),
@@ -206,7 +206,7 @@ const struct SpriteFrameImage gEventObjectPicTable_HondewBerryTree[] = {
     overworld_frame(gEventObjectPic_HondewBerryTree, 2, 4, 5),
 };
 
-const u8 gBerryTreePaletteSlotTable_Hondew[] = {3, 4, 5, 5, 5};
+const u16 gBerryTreePaletteSlotTable_Hondew[] = {3, 4, 5, 5, 5};
 
 const struct SpriteFrameImage gEventObjectPicTable_AguavBerryTree[] = {
     overworld_frame(gEventObjectPic_BerryTreeDirtPile, 2, 2, 0),
@@ -220,7 +220,7 @@ const struct SpriteFrameImage gEventObjectPicTable_AguavBerryTree[] = {
     overworld_frame(gEventObjectPic_AguavBerryTree, 2, 4, 5),
 };
 
-const u8 gBerryTreePaletteSlotTable_Aguav[] = {3, 4, 4, 4, 4};
+const u16 gBerryTreePaletteSlotTable_Aguav[] = {3, 4, 4, 4, 4};
 
 const struct SpriteFrameImage gEventObjectPicTable_WikiBerryTree[] = {
     overworld_frame(gEventObjectPic_BerryTreeDirtPile, 2, 2, 0),
@@ -234,7 +234,7 @@ const struct SpriteFrameImage gEventObjectPicTable_WikiBerryTree[] = {
     overworld_frame(gEventObjectPic_WikiBerryTree, 2, 4, 5),
 };
 
-const u8 gBerryTreePaletteSlotTable_Wiki[] = {3, 4, 3, 3, 3};
+const u16 gBerryTreePaletteSlotTable_Wiki[] = {3, 4, 3, 3, 3};
 
 const struct SpriteFrameImage gEventObjectPicTable_PomegBerryTree[] = {
     overworld_frame(gEventObjectPic_BerryTreeDirtPile, 2, 2, 0),
@@ -248,7 +248,7 @@ const struct SpriteFrameImage gEventObjectPicTable_PomegBerryTree[] = {
     overworld_frame(gEventObjectPic_PomegBerryTree, 2, 4, 5),
 };
 
-const u8 gBerryTreePaletteSlotTable_Pomeg[] = {3, 4, 3, 3, 3};
+const u16 gBerryTreePaletteSlotTable_Pomeg[] = {3, 4, 3, 3, 3};
 
 const struct SpriteFrameImage gEventObjectPicTable_RawstBerryTree[] = {
     overworld_frame(gEventObjectPic_BerryTreeDirtPile, 2, 2, 0),
@@ -262,7 +262,7 @@ const struct SpriteFrameImage gEventObjectPicTable_RawstBerryTree[] = {
     overworld_frame(gEventObjectPic_RawstBerryTree, 2, 4, 5),
 };
 
-const u8 gBerryTreePaletteSlotTable_Rawst[] = {3, 4, 4, 4, 4};
+const u16 gBerryTreePaletteSlotTable_Rawst[] = {3, 4, 4, 4, 4};
 
 const struct SpriteFrameImage gEventObjectPicTable_SpelonBerryTree[] = {
     overworld_frame(gEventObjectPic_BerryTreeDirtPile, 2, 2, 0),
@@ -276,7 +276,7 @@ const struct SpriteFrameImage gEventObjectPicTable_SpelonBerryTree[] = {
     overworld_frame(gEventObjectPic_SpelonBerryTree, 2, 4, 5),
 };
 
-const u8 gBerryTreePaletteSlotTable_Spelon[] = {3, 4, 3, 3, 3};
+const u16 gBerryTreePaletteSlotTable_Spelon[] = {3, 4, 3, 3, 3};
 
 const struct SpriteFrameImage gEventObjectPicTable_ChestoBerryTree[] = {
     overworld_frame(gEventObjectPic_BerryTreeDirtPile, 2, 2, 0),
@@ -290,7 +290,7 @@ const struct SpriteFrameImage gEventObjectPicTable_ChestoBerryTree[] = {
     overworld_frame(gEventObjectPic_ChestoBerryTree, 2, 4, 5),
 };
 
-const u8 gBerryTreePaletteSlotTable_Chesto[] = {3, 4, 2, 2, 2};
+const u16 gBerryTreePaletteSlotTable_Chesto[] = {3, 4, 2, 2, 2};
 
 const struct SpriteFrameImage gEventObjectPicTable_OranBerryTree[] = {
     overworld_frame(gEventObjectPic_BerryTreeDirtPile, 2, 2, 0),
@@ -304,7 +304,7 @@ const struct SpriteFrameImage gEventObjectPicTable_OranBerryTree[] = {
     overworld_frame(gEventObjectPic_OranBerryTree, 2, 4, 5),
 };
 
-const u8 gBerryTreePaletteSlotTable_Oran[] = {3, 4, 2, 2, 2};
+const u16 gBerryTreePaletteSlotTable_Oran[] = {3, 4, 2, 2, 2};
 
 const struct SpriteFrameImage gEventObjectPicTable_PersimBerryTree[] = {
     overworld_frame(gEventObjectPic_BerryTreeDirtPile, 2, 2, 0),
@@ -318,7 +318,7 @@ const struct SpriteFrameImage gEventObjectPicTable_PersimBerryTree[] = {
     overworld_frame(gEventObjectPic_PersimBerryTree, 2, 4, 5),
 };
 
-const u8 gBerryTreePaletteSlotTable_Persim[] = {3, 4, 2, 2, 2};
+const u16 gBerryTreePaletteSlotTable_Persim[] = {3, 4, 2, 2, 2};
 
 const struct SpriteFrameImage gEventObjectPicTable_SitrusBerryTree[] = {
     overworld_frame(gEventObjectPic_BerryTreeDirtPile, 2, 2, 0),
@@ -332,7 +332,7 @@ const struct SpriteFrameImage gEventObjectPicTable_SitrusBerryTree[] = {
     overworld_frame(gEventObjectPic_SitrusBerryTree, 2, 4, 5),
 };
 
-const u8 gBerryTreePaletteSlotTable_Sitrus[] = {3, 4, 4, 4, 4};
+const u16 gBerryTreePaletteSlotTable_Sitrus[] = {3, 4, 4, 4, 4};
 
 const struct SpriteFrameImage gEventObjectPicTable_AspearBerryTree[] = {
     overworld_frame(gEventObjectPic_BerryTreeDirtPile, 2, 2, 0),
@@ -346,7 +346,7 @@ const struct SpriteFrameImage gEventObjectPicTable_AspearBerryTree[] = {
     overworld_frame(gEventObjectPic_AspearBerryTree, 2, 4, 5),
 };
 
-const u8 gBerryTreePaletteSlotTable_Aspear[] = {3, 4, 3, 3, 3};
+const u16 gBerryTreePaletteSlotTable_Aspear[] = {3, 4, 3, 3, 3};
 
 const struct SpriteFrameImage gEventObjectPicTable_PamtreBerryTree[] = {
     overworld_frame(gEventObjectPic_BerryTreeDirtPile, 2, 2, 0),
@@ -360,7 +360,7 @@ const struct SpriteFrameImage gEventObjectPicTable_PamtreBerryTree[] = {
     overworld_frame(gEventObjectPic_PamtreBerryTree, 2, 4, 5),
 };
 
-const u8 gBerryTreePaletteSlotTable_Pamtre[] = {3, 4, 2, 2, 2};
+const u16 gBerryTreePaletteSlotTable_Pamtre[] = {3, 4, 2, 2, 2};
 
 const struct SpriteFrameImage gEventObjectPicTable_CornnBerryTree[] = {
     overworld_frame(gEventObjectPic_BerryTreeDirtPile, 2, 2, 0),
@@ -374,7 +374,7 @@ const struct SpriteFrameImage gEventObjectPicTable_CornnBerryTree[] = {
     overworld_frame(gEventObjectPic_CornnBerryTree, 2, 4, 5),
 };
 
-const u8 gBerryTreePaletteSlotTable_Cornn[] = {3, 4, 2, 2, 2};
+const u16 gBerryTreePaletteSlotTable_Cornn[] = {3, 4, 2, 2, 2};
 
 const struct SpriteFrameImage gEventObjectPicTable_LansatBerryTree[] = {
     overworld_frame(gEventObjectPic_BerryTreeDirtPile, 2, 2, 0),
@@ -388,7 +388,7 @@ const struct SpriteFrameImage gEventObjectPicTable_LansatBerryTree[] = {
     overworld_frame(gEventObjectPic_LansatBerryTree, 2, 4, 5),
 };
 
-const u8 gBerryTreePaletteSlotTable_Lansat[] = {3, 4, 2, 2, 2};
+const u16 gBerryTreePaletteSlotTable_Lansat[] = {3, 4, 2, 2, 2};
 
 const struct SpriteFrameImage gEventObjectPicTable_DurinBerryTree[] = {
     overworld_frame(gEventObjectPic_BerryTreeDirtPile, 2, 2, 0),
@@ -402,7 +402,7 @@ const struct SpriteFrameImage gEventObjectPicTable_DurinBerryTree[] = {
     overworld_frame(gEventObjectPic_DurinBerryTree, 2, 4, 5),
 };
 
-const u8 gBerryTreePaletteSlotTable_Durin[] = {3, 4, 4, 4, 4};
+const u16 gBerryTreePaletteSlotTable_Durin[] = {3, 4, 4, 4, 4};
 
 const struct SpriteFrameImage gEventObjectPicTable_TamatoBerryTree[] = {
     overworld_frame(gEventObjectPic_BerryTreeDirtPile, 2, 2, 0),
@@ -416,7 +416,7 @@ const struct SpriteFrameImage gEventObjectPicTable_TamatoBerryTree[] = {
     overworld_frame(gEventObjectPic_TamatoBerryTree, 2, 4, 5),
 };
 
-const u8 gBerryTreePaletteSlotTable_Tamato[] = {3, 4, 2, 2, 2};
+const u16 gBerryTreePaletteSlotTable_Tamato[] = {3, 4, 2, 2, 2};
 
 const struct SpriteFrameImage gEventObjectPicTable_OccaBerryTree[] = {
     overworld_frame(gEventObjectPic_BerryTreeDirtPile, 2, 2, 0),
@@ -430,7 +430,7 @@ const struct SpriteFrameImage gEventObjectPicTable_OccaBerryTree[] = {
     overworld_frame(gEventObjectPic_OccaBerryTree, 2, 4, 5)
 };
 
-const u8 gBerryTreePaletteSlotTable_Occa[] = {3, 4, 4, 4, 4};
+const u16 gBerryTreePaletteSlotTable_Occa[] = {3, 4, 4, 4, 4};
 
 const struct SpriteFrameImage gEventObjectPicTable_YacheBerryTree[] = {
     overworld_frame(gEventObjectPic_BerryTreeDirtPile, 2, 2, 0),
@@ -444,7 +444,7 @@ const struct SpriteFrameImage gEventObjectPicTable_YacheBerryTree[] = {
     overworld_frame(gEventObjectPic_YacheBerryTree, 2, 4, 5)
 };
 
-const u8 gBerryTreePaletteSlotTable_Yache[] = {3, 4, 4, 4, 4};
+const u16 gBerryTreePaletteSlotTable_Yache[] = {3, 4, 4, 4, 4};
 
 const struct SpriteFrameImage gEventObjectPicTable_ChopleBerryTree[] = {
     overworld_frame(gEventObjectPic_BerryTreeDirtPile, 2, 2, 0),
@@ -458,7 +458,7 @@ const struct SpriteFrameImage gEventObjectPicTable_ChopleBerryTree[] = {
     overworld_frame(gEventObjectPic_ChopleBerryTree, 2, 4, 5)
 };
 
-const u8 gBerryTreePaletteSlotTable_Chople[] = {3, 4, 3, 3, 3};
+const u16 gBerryTreePaletteSlotTable_Chople[] = {3, 4, 3, 3, 3};
 
 const struct SpriteFrameImage gEventObjectPicTable_KebiaBerryTree[] = {
     overworld_frame(gEventObjectPic_BerryTreeDirtPile, 2, 2, 0),
@@ -472,7 +472,7 @@ const struct SpriteFrameImage gEventObjectPicTable_KebiaBerryTree[] = {
     overworld_frame(gEventObjectPic_KebiaBerryTree, 2, 4, 5)
 };
 
-const u8 gBerryTreePaletteSlotTable_Kebia[] = {3, 4, 4, 4, 4};
+const u16 gBerryTreePaletteSlotTable_Kebia[] = {3, 4, 4, 4, 4};
 
 const struct SpriteFrameImage gEventObjectPicTable_ShucaBerryTree[] = {
     overworld_frame(gEventObjectPic_BerryTreeDirtPile, 2, 2, 0),
@@ -486,7 +486,7 @@ const struct SpriteFrameImage gEventObjectPicTable_ShucaBerryTree[] = {
     overworld_frame(gEventObjectPic_ShucaBerryTree, 2, 4, 5)
 };
 
-const u8 gBerryTreePaletteSlotTable_Shuca[] = {3, 4, 5, 5, 5};
+const u16 gBerryTreePaletteSlotTable_Shuca[] = {3, 4, 5, 5, 5};
 
 const struct SpriteFrameImage gEventObjectPicTable_PayapaBerryTree[] = {
     overworld_frame(gEventObjectPic_BerryTreeDirtPile, 2, 2, 0),
@@ -500,7 +500,7 @@ const struct SpriteFrameImage gEventObjectPicTable_PayapaBerryTree[] = {
     overworld_frame(gEventObjectPic_PayapaBerryTree, 2, 4, 5)
 };
 
-const u8 gBerryTreePaletteSlotTable_Payapa[] = {3, 4, 5, 5, 5};
+const u16 gBerryTreePaletteSlotTable_Payapa[] = {3, 4, 5, 5, 5};
 
 const struct SpriteFrameImage gEventObjectPicTable_TangaBerryTree[] = {
     overworld_frame(gEventObjectPic_BerryTreeDirtPile, 2, 2, 0),
@@ -514,7 +514,7 @@ const struct SpriteFrameImage gEventObjectPicTable_TangaBerryTree[] = {
     overworld_frame(gEventObjectPic_TangaBerryTree, 2, 4, 5)
 };
 
-const u8 gBerryTreePaletteSlotTable_Tanga[] = {3, 4, 4, 4, 4};
+const u16 gBerryTreePaletteSlotTable_Tanga[] = {3, 4, 4, 4, 4};
 
 const struct SpriteFrameImage gEventObjectPicTable_KasibBerryTree[] = {
     overworld_frame(gEventObjectPic_BerryTreeDirtPile, 2, 2, 0),
@@ -528,7 +528,7 @@ const struct SpriteFrameImage gEventObjectPicTable_KasibBerryTree[] = {
     overworld_frame(gEventObjectPic_KasibBerryTree, 2, 4, 5)
 };
 
-const u8 gBerryTreePaletteSlotTable_Kasib[] = {3, 4, 4, 4, 4};
+const u16 gBerryTreePaletteSlotTable_Kasib[] = {3, 4, 4, 4, 4};
 
 const struct SpriteFrameImage gEventObjectPicTable_HabanBerryTree[] = {
     overworld_frame(gEventObjectPic_BerryTreeDirtPile, 2, 2, 0),
@@ -542,7 +542,7 @@ const struct SpriteFrameImage gEventObjectPicTable_HabanBerryTree[] = {
     overworld_frame(gEventObjectPic_HabanBerryTree, 2, 4, 5)
 };
 
-const u8 gBerryTreePaletteSlotTable_Haban[] = {3, 4, 3, 3, 3};
+const u16 gBerryTreePaletteSlotTable_Haban[] = {3, 4, 3, 3, 3};
 
 const struct SpriteFrameImage gEventObjectPicTable_ColburBerryTree[] = {
     overworld_frame(gEventObjectPic_BerryTreeDirtPile, 2, 2, 0),
@@ -556,7 +556,7 @@ const struct SpriteFrameImage gEventObjectPicTable_ColburBerryTree[] = {
     overworld_frame(gEventObjectPic_ColburBerryTree, 2, 4, 5)
 };
 
-const u8 gBerryTreePaletteSlotTable_Colbur[] = {3, 4, 5, 5, 5};
+const u16 gBerryTreePaletteSlotTable_Colbur[] = {3, 4, 5, 5, 5};
 
 const struct SpriteFrameImage gEventObjectPicTable_RoseliBerryTree[] = {
     overworld_frame(gEventObjectPic_BerryTreeDirtPile, 2, 2, 0),
@@ -570,11 +570,11 @@ const struct SpriteFrameImage gEventObjectPicTable_RoseliBerryTree[] = {
     overworld_frame(gEventObjectPic_RoseliBerryTree, 2, 4, 5)
 };
 
-const u8 gBerryTreePaletteSlotTable_Roseli[] = {3, 4, 4, 4, 4};
+const u16 gBerryTreePaletteSlotTable_Roseli[] = {3, 4, 4, 4, 4};
 
-const u8 gDeadBerryTreeEventObjectGraphicsIdTable[] = {EVENT_OBJ_GFX_BERRY_TREE_EARLY_STAGES, EVENT_OBJ_GFX_BERRY_TREE_EARLY_STAGES, EVENT_OBJ_GFX_BERRY_TREE_EARLY_STAGES, EVENT_OBJ_GFX_BERRY_TREE_EARLY_STAGES, EVENT_OBJ_GFX_BERRY_TREE_EARLY_STAGES};
+const u16 gDeadBerryTreeEventObjectGraphicsIdTable[] = {EVENT_OBJ_GFX_BERRY_TREE_EARLY_STAGES, EVENT_OBJ_GFX_BERRY_TREE_EARLY_STAGES, EVENT_OBJ_GFX_BERRY_TREE_EARLY_STAGES, EVENT_OBJ_GFX_BERRY_TREE_EARLY_STAGES, EVENT_OBJ_GFX_BERRY_TREE_EARLY_STAGES};
 
-const u8 gBerryTreeEventObjectGraphicsIdTable[] = {EVENT_OBJ_GFX_BERRY_TREE_EARLY_STAGES, EVENT_OBJ_GFX_BERRY_TREE_EARLY_STAGES, EVENT_OBJ_GFX_BERRY_TREE_LATE_STAGES, EVENT_OBJ_GFX_BERRY_TREE_LATE_STAGES, EVENT_OBJ_GFX_BERRY_TREE_LATE_STAGES};
+const u16 gBerryTreeEventObjectGraphicsIdTable[] = {EVENT_OBJ_GFX_BERRY_TREE_EARLY_STAGES, EVENT_OBJ_GFX_BERRY_TREE_EARLY_STAGES, EVENT_OBJ_GFX_BERRY_TREE_LATE_STAGES, EVENT_OBJ_GFX_BERRY_TREE_LATE_STAGES, EVENT_OBJ_GFX_BERRY_TREE_LATE_STAGES};
 
 const struct SpriteFrameImage *const gBerryTreePicTablePointers[] = {
     [ITEM_CHERI_BERRY - FIRST_BERRY_INDEX] = gEventObjectPicTable_CheriBerryTree,
@@ -646,7 +646,7 @@ const struct SpriteFrameImage *const gBerryTreePicTablePointers[] = {
     [ITEM_MARANGA_BERRY - FIRST_BERRY_INDEX] = NULL,
 };
 
-const u8 *const gBerryTreePaletteSlotTablePointers[] = {
+const u16 *const gBerryTreePaletteSlotTablePointers[] = {
     [ITEM_CHERI_BERRY - FIRST_BERRY_INDEX] = gBerryTreePaletteSlotTable_Cheri,
     [ITEM_CHESTO_BERRY - FIRST_BERRY_INDEX] = gBerryTreePaletteSlotTable_Chesto,
     [ITEM_PECHA_BERRY - FIRST_BERRY_INDEX] = gBerryTreePaletteSlotTable_Pecha,
@@ -716,7 +716,7 @@ const u8 *const gBerryTreePaletteSlotTablePointers[] = {
     [ITEM_MARANGA_BERRY - FIRST_BERRY_INDEX] = NULL,
 };
 
-const u8 *const gBerryTreeEventObjectGraphicsIdTablePointers[] = {
+const u16 *const gBerryTreeEventObjectGraphicsIdTablePointers[] = {
     [ITEM_CHERI_BERRY - FIRST_BERRY_INDEX] = gBerryTreeEventObjectGraphicsIdTable,
     [ITEM_CHESTO_BERRY - FIRST_BERRY_INDEX] = gBerryTreeEventObjectGraphicsIdTable,
     [ITEM_PECHA_BERRY - FIRST_BERRY_INDEX] = gBerryTreeEventObjectGraphicsIdTable,
